+++ src/js/integrations/index.ts
@@ -1,3 +1,4 @@
 export { Cordova } from './cordova';
 export { EventOrigin } from './eventorigin';
 export { SdkInfo } from './sdkinfo';
+export { ReplayBrowserFilter } from './replayBrowserFilter';